+++ src/main/java/net/maplecraft/utils/MapleCraftConstants.java
@@ -1,5 +1,5 @@
 package net.maplecraft.utils;
 
 public class MapleCraftConstants {
-    public static final int MAX_PLAYER_MANA_POINTS = 20;
+    public static final double MAX_PLAYER_MANA_POINTS = 20.0D;
 }
